pyproject.toml
[project]
name = "webtools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "gunicorn>=23.0.0",
    "pillow>=11.3.0",
]
